Add files via upload
The entirety of the python, HTML / CSS code for OncoDetect. 'OncoDetect.py' contains the code for the neural network that powers this AI model, 'ondodetectFlask.py' contains the code for the web application that allows users to upload medical images, 'index.html, style.css' contain the HTML and CSS code for the front end of the application. The flask application acts as the middleman between the user and the neural network as It makes this application very easy to use and also allows the application to accept any medical images the user may want to upload. NOTE: this model was trained using a breast cancer dataset ... this dataset contained roughly 500 images and is not uploaded to GitHub. I only uploaded my code for the application.